check if on Arch before installing initcpio hook
This commit modifies the `Makefile` and adds  a condition that checks if
`mkinitcpio` is  available in the path,  and only then the initcpio hook
will be installed.

On distros other than Arch and the likes,  this prevents two 'rogue' fi-
les from being installed to a system location without having any use.