[Makefile] remove `mkinitcpio` dependency
The mkinitcpio binary is not available even on Arch,  when building in a
clean environment (e.g. building distro packages in a clean chroot).

This causes the mkinitcpio hook  to not be present in the final package,
which is resolved by this commit, by adding a `make` switch. This switch
allows users to explicitely turn on inclusion of the hook.

Ref.:
https://github.com/Antynea/grub-btrfs/pull/135#issuecomment-761806829