[make help] mention `INITCPIO` switch
This commit is an expansion to :
https://github.com/Antynea/grub-btrfs/commit/87b816345bd790b283336c693dcc02862149b6b7

Since a new `make` parameter has been introduced, it should be mentioned
in the output of the `help` target.

Also, the table is being expanded by a new column describing the type of
data that each parameter may be assigned, like paths or boolean values.